DO NOT USE! How to fork from freicoin foundation
DO NOT USE!

example of how to fork off of the freicoin foundation to create a new coin without having to restart the entire block chain. 
There are two methods to this, the one is simply stating a block that is agreed upon to stop issuing new coins.

the other method (which is included here) gives the foundation a bit of time for some more coins, then the rest meant for a foundation would be tossed to an eater / black  hole address. Those coins would (over the course of 20 years) demurrage back into the coinbase.
The foundation still has some coins, but not as many as were designed at the start. Since there is a possibility of coins getting demurraged anyway, lets just make it happen instead of wasting time giving out more coins that are going to do nothing but sit in someone's hands.

DO NOT USE!